requirements.txt
websockets==8.1
aiohttp==3.6.2
discord.py==1.2.5
